Web/BE/requirements.txt
asgiref==3.8.1
boto3==1.34.149
botocore==1.34.149
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==4.2.14
django-allauth==0.63.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-mysql==4.14.0
django-storages==1.14.4
djangorestframework==3.15.2
Faker==26.0.0
idna==3.7
jmespath==1.0.1
mysqlclient==2.2.4
paho-mqtt==2.1.0
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
